feat: reuse profile card in profile header
